Moved products dialog to settings
