+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/WooPosHomeState.kt
@@ -1,7 +1,6 @@
 package com.woocommerce.android.ui.woopos.home
 
 import android.os.Parcelable
-import androidx.annotation.StringRes
 import com.woocommerce.android.R
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
@@ -30,30 +29,6 @@ data class WooPosHomeState(
     sealed class DialogState : Parcelable {
         @Parcelize
         data object Hidden : DialogState()
-
-        @Parcelize
-        data object ProductsInfoDialog : DialogState() {
-            @IgnoredOnParcel
-            val header: Int = R.string.woopos_dialog_products_info_heading
-
-            @IgnoredOnParcel
-            val primaryMessage: Int = R.string.woopos_dialog_products_info_primary_message
-
-            @IgnoredOnParcel
-            val secondaryMessage: Int = R.string.woopos_dialog_products_info_secondary_message
-
-            @IgnoredOnParcel
-            val tertiaryMessage: Int = R.string.woopos_dialog_products_info_tertiary_message
-
-            @IgnoredOnParcel
-            val primaryButton: PrimaryButton = PrimaryButton(
-                label = R.string.woopos_dialog_products_info_button_label,
-            )
-
-            data class PrimaryButton(
-                @StringRes val label: Int,
-            )
-        }
 
         @Parcelize
         data object ScanningSetupDialog : DialogState()
